+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh3w7zbc7q8rzhg6x9z8x34f"
+        "conv_01kh3yn5rfr141knw458k7spmc"
     )
     r"""Unique conversation identifier with `conv_` prefix."""