+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh3w7zbe01xw7hhmw7xzbwmm"
+        "conv_01kh3yn5rhp43xjjz6smvvha45"
     )
     r"""Unique conversation identifier with `conv_` prefix."""