+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-10T13:35:42.272Z")
+    created: Optional[datetime] = parse_datetime("2026-02-10T14:17:51.632Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")