+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-10T13:35:38.743Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-10T14:17:47.971Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")